Changes needed for running in docker
* README.md - added an explanation on how to uild the docker image
* webpack.config.js - set host to 0.0.0.0 so that the dev server was
available externally to the docker instance
* Dockerfile - added new docker file

This docker file uses "yarn start" to run the image.  This may not be
how we wish to do this in production but will allow us to proceed to
testing. Prior to production we should discuss how to run the
application in a production ready manner.